# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShopOS Auto Repair Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue, void]
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- REPAIR ORDERS ----
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemInput"
      responses:
        "201":
          description: Inventory item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemInput"
      responses:
        "200":
          description: Inventory item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- INSPECTIONS ----
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: invoiceId
          in: query
          schema:
            type: integer
        - name: method
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- EMPLOYEES ----
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        vehicleCount:
          type: integer
        totalBilled:
          type: number
        totalPaid:
          type: number
      required:
        - id
        - firstName
        - lastName
        - createdAt
        - updatedAt
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        vin:
          type: string
        licensePlate:
          type: string
        year:
          type: integer
        make:
          type: string
        model:
          type: string
        trim:
          type: string
        color:
          type: string
        mileage:
          type: integer
        engineType:
          type: string
        transmissionType:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customer:
          $ref: "#/components/schemas/Customer"
      required:
        - id
        - customerId
        - year
        - make
        - model
        - createdAt
        - updatedAt
...
    LineItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [labor, part, fee, discount]
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
        partNumber:
          type: string
        inventoryItemId:
          type: integer
      required:
        - type
        - description
        - quantity
        - unitPrice
        - total
    CreateLineItemInput:
      type: object
      properties:
        type:
          type: string
          enum: [labor, part, fee, discount]
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        partNumber:
          type: string
        inventoryItemId:
          type: integer
      required:
        - type
        - description
        - quantity
        - unitPrice
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        vehicleId:
          type: integer
        repairOrderId:
          type: integer
        estimateId:
          type: integer
        status:
          type: string
          enum: [draft, sent, paid, overdue, void]
        notes:
          type: string
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        dueDate:
          type: string
          format: date
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        customer:
          $ref: "#/components/schemas/Customer"
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - customerId
        - status
        - subtotal
        - taxRate
        - taxAmount
        - discountAmount
        - total
        - amountPaid
        - balance
        - createdAt
        - updatedAt
    CreateInvoiceInput:
      type: object
      properties:
        customerId:
          type: integer
        vehicleId:
          type: integer
        repairOrderId:
          type: integer
        estimateId:
          type: integer
        status:
          type: string
          enum: [draft, sent, paid, overdue, void]
        notes:
          type: string
        taxRate:
          type: number
        discountAmount:
          type: number
        dueDate:
          type: string
          format: date
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateLineItemInput"
      required:
        - customerId
        - lineItems
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        partNumber:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        vendor:
          type: string
        costPrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        minQuantity:
          type: integer
        location:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - costPrice
        - sellPrice
        - quantity
        - minQuantity
        - createdAt
        - updatedAt
    CreateInventoryItemInput:
      type: object
      properties:
        partNumber:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        vendor:
          type: string
        costPrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        minQuantity:
          type: integer
        location:
          type: string
        notes:
          type: string
      required:
        - name
        - category
        - costPrice
        - sellPrice
        - quantity
        - minQuantity
    InventoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, check, credit_card, debit_card, paypal, square, stripe, other]
        referenceNumber:
          type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceId
        - amount
        - method
        - paidAt
        - createdAt
    CreatePaymentInput:
      type: object
      properties:
        invoiceId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, check, credit_card, debit_card, paypal, square, stripe, other]
        referenceNumber:
          type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
      required:
        - invoiceId
        - amount
        - method
    PaymentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...